Move yet another chunk of logic
