+++ packages/truffle-compile/profiler/index.js
@@ -3,7 +3,6 @@
 
 const path = require("path");
 const async = require("async");
-const fse = require("fs-extra");
 const Parser = require("../parser");
 const CompilerSupplier = require("../compilerSupplier");
 const expect = require("truffle-expect");
@@ -12,6 +11,7 @@ const semver = require("semver");
 const debug = require("debug")("compile:profiler");
 const readAndParseArtifactFiles = require("./readAndParseArtifactFiles");
 const minimumUpdatedTimePerSource = require("./minimumUpdatedTimePerSource");
+const findUpdatedFiles = require("./findUpdatedFiles");
 
 module.exports = {
   updated(options, callback) {
@@ -30,7 +30,7 @@ module.exports = {
     let sourceFilesArtifacts = {};
     let sourceFilesArtifactsUpdatedTimes = {};
 
-    const updatedFiles = [];
+    let updatedFiles = [];
 
     getFiles()
       .then(sourceFiles => {
@@ -47,38 +47,10 @@ module.exports = {
         return;
       })
       .then(() => {
-        // Stat all the source files, getting there updated times, and comparing them to
-        // the artifact updated times.
-        const sourceFiles = Object.keys(sourceFilesArtifacts);
-
-        let sourceFileStats;
-        sourceFileStats = sourceFiles.map(file => {
-          try {
-            return fse.statSync(file);
-          } catch (error) {
-            // Ignore it. This means the source file was removed
-            // but the artifact file possibly exists. Return null
-            // to signfy that we should ignore it.
-            return null;
-          }
-        });
-
-        sourceFiles.forEach((sourceFile, index) => {
-          const sourceFileStat = sourceFileStats[index];
-
-          // Ignore updating artifacts if source file has been removed.
-          if (sourceFileStat == null) return;
-
-          const artifactsUpdatedTime =
-            sourceFilesArtifactsUpdatedTimes[sourceFile] || 0;
-          const sourceFileUpdatedTime = (
-            sourceFileStat.mtime || sourceFileStat.ctime
-          ).getTime();
-
-          if (sourceFileUpdatedTime > artifactsUpdatedTime) {
-            updatedFiles.push(sourceFile);
-          }
-        });
+        updatedFiles = findUpdatedFiles(
+          sourceFilesArtifacts,
+          sourceFilesArtifactsUpdatedTimes
+        );
         return callback(null, updatedFiles);
       })
       .catch(error => {
